+++ src/Plivo/Resource/Call/AsyncResponse.cs
@@ -1,0 +1,6 @@
+namespace Plivo.Resource.Call
+{
+    public class AsyncResponse : BaseResponse
+    {
+    }
+}