+++ src/Plivo/Resource/Recording/AsyncResponse.cs
@@ -1,0 +1,6 @@
+namespace Plivo.Resource.Recording
+{
+    public class AsyncResponse : BaseResponse
+    {
+    }
+}